+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Both mode:
 Client mode:
 * Issue 65: When using requests with the parameters in the request (no GraphQL Variables and no Bind Parameter), the request is properly encoded when these parameters are or contain strings
 * Dependency order changed in the graphql-java-client-dependencies module, to make sure the right spring's dependencies are used (this could prevent a Spring app to start)
+* Removed the use of the `reactor.core.publisher.Sinks` class, to maintain compatibility for those who uses an older version of Spring Boot
 
 # 1.14